test(amber): split WorkerManagersAdditionalSpec into per-source-class files
Address Yicong-Huang's review on #5439: the single
WorkerManagersAdditionalSpec.scala combining tests for two source
classes makes future contributors miss this file when they search
StatisticsManagerSpec.scala / SerializationManagerSpec.scala — exactly
the duplicate-spec failure mode that surfaced on #5400 with
CoreExecutorReflectionSpec vs my parallel CoreExecutorSpec.

Split into two files following the <srcClassName>Spec.scala convention:
- StatisticsManagerSpec.scala (12 tests for StatisticsManager)
- SerializationManagerSpec.scala (8 tests for SerializationManager)

No assertion changes; tests + companion fixtures simply move to the
file that matches their source class.
